refactor visualization functions
